qchat-web/src/backend/requirements.txt
#for imports
azure-functions
langchain
langchain-core
langchain-ollama
bs4
faiss-cpu>=1.9.0
httpx
requests
pymongo 
certifi
python-dotenv
beautifulsoup4==4.12.3
langchain-ollama==0.1.2
requests==2.32.3
